Addressed PR changes.
Separated and renamed breadcrumb components.
Merged develop.
Removed Home link from breadcrumb nav.
