+++ src/helpers/types.ts
@@ -1,4 +1,4 @@
-import { Breadcrumb } from '../components/Breadcrumbs'
+import { Breadcrumb } from '../components/Breadcrumbs/Breadcrumbs'
 
 export type ArrayElement<A> = A extends readonly (infer T)[] ? T : never
